move EncryptedNote to types
